backend/requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.22.0
neo4j==5.8.1
redis==4.5.5
pydantic==1.10.7
httpx==0.24.0
openai==0.27.6
python-dotenv==1.0.0
aiohttp==3.8.4